src/impl_: prepare ModuleDef for per-module state init
This commit sets ModuleDef.m_size to size_of::<ModuleState>, and adds
the relevant callback to ModuleDef.m_free.

We also add C function defs for Py_mod_exec and ModuleDef.m_free
callbacks. These will be used to initialize and free ModuleState in
later patches.

We do not yet set an initializer, as these are passed in m_slots, which
are set from the `#[pymodule]` impl in pyo3-macros-backend.